% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint_analysis.R
\name{conjoint_analysis}
\alias{conjoint_analysis}
\title{Title Perfom Conjoint Analysis}
\usage{
conjoint_analysis(y, x, z, y.type = "score")
}
\arguments{
\item{y.type}{}
}
\value{

}
\description{
Title Perfom Conjoint Analysis
}
